Sftp: rebase Signed-off-by: Brendan Cronin <brendan.cronin@mercedes-benz.com>
