--- edc-extensions/transferprocess-sftp-client/src/main/java/net/catenax/edc/transferprocess/sftp/client/SftpClientExtension.java
@@ -1,28 +1,0 @@
-/*
- *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
- *
- */
-
-package net.catenax.edc.transferprocess.sftp.client;
-
-import org.eclipse.dataspaceconnector.runtime.metamodel.annotation.Provides;
-import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
-import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-
-@Provides(SftpClientWrapper.class)
-public class SftpClientExtension implements ServiceExtension {
-    @Override
-    public void initialize(ServiceExtensionContext context) {
-        SshdSftpClient sftpClient = new SshdSftpClient();
-        context.registerService(SftpClientWrapper.class, sftpClient);
-    }
-}
