--- edc-extensions/transferprocess-sftp-common/src/main/java/net/catenax/edc/trasnferprocess/sftp/common/SftpLocation.java
@@ -1,30 +1,0 @@
-/*
- *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
- *
- */
-
-package net.catenax.edc.trasnferprocess.sftp.common;
-
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-@Builder
-@Getter
-@ToString
-@EqualsAndHashCode
-public class SftpLocation {
-    private final String host;
-    private final Integer port;
-    private final String path;
-}
