--- edc-extensions/transferprocess-sftp-common/src/main/java/net/catenax/edc/trasnferprocess/sftp/common/SftpLocationFactory.java
@@ -1,19 +1,0 @@
-/*
- *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
- *
- */
-
-package net.catenax.edc.trasnferprocess.sftp.common;
-
-public interface SftpLocationFactory {
-    SftpLocation createSftpLocation(String transferProcessId);
-}
