--- edc-extensions/transferprocess-sftp-common/src/main/java/net/catenax/edc/trasnferprocess/sftp/common/SftpUser.java
@@ -1,32 +1,0 @@
-/*
- *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
- *
- */
-
-package net.catenax.edc.trasnferprocess.sftp.common;
-
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-import java.security.KeyPair;
-
-@Builder
-@Getter
-@ToString(of = "name")
-@EqualsAndHashCode
-public class SftpUser {
-    private final String name;
-    private final String password;
-    private final KeyPair keyPair;
-}
