--- edc-extensions/transferprocess-sftp-provisioner/src/main/java/net/catenax/edc/transferprocess/sftp/provisioner/SftpProviderResourceDefinition.java
@@ -1,32 +1,0 @@
-/*
- *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
- *
- */
-
-package net.catenax.edc.transferprocess.sftp.provisioner;
-
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ResourceDefinition;
-
-@Getter
-@RequiredArgsConstructor
-public class SftpProviderResourceDefinition extends ResourceDefinition {
-    @NonNull
-    private String dataAddressType;
-
-    @Override
-    public <RD extends ResourceDefinition, B extends Builder<RD, B>> B toBuilder() {
-        return null;
-    }
-}
