+++ edc-extensions/transferprocess-sftp-provisioner/src/main/java/org/eclipse/tractusx/edc/transferprocess/sftp/provisioner/ConfigBackedSftpLocationFactory.java
@@ -1,0 +1,35 @@
+/*
+ *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
+ *
+ */
+
+package org.eclipse.tractusx.edc.transferprocess.sftp.provisioner;
+
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpLocation;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpLocationFactory;
+
+@RequiredArgsConstructor
+public class ConfigBackedSftpLocationFactory implements SftpLocationFactory {
+    @NonNull
+    private final String sftpHost;
+    @NonNull
+    private final Integer sftpPort;
+    @NonNull
+    private final String sftpPath;
+
+    @Override
+    public SftpLocation createSftpLocation(String transferProcessId) {
+        return SftpLocation.builder().host(sftpHost).port(sftpPort).path(sftpPath).build();
+    }
+}
